MSE-in-Workers: Begin to spec cross-context track object aliasing
AudioTrack, VideoTrack, TextTrack extensions' sourceBuffer attribute
must be null in the Window context's alias of the track if the track
is an alias to a worker-created track from MSE-in-Worker. This
is to prevent any assumption that cross-context object references
are somehow now allowed in MSE or HTMLMediaElement+MSE. More spec
language is necessary to describe the creation and management of
these track aliases. There also needs to be description of extension
of {Audio,Video,Track}{,List} IDL to be exposed in DedicatedWorker. This
latter item might need to be directly in the media element spec; may
file issue to discuss in interim.